+++ src/testFixtures/java/argo/ElementTrackingStajParserMatcher.java
@@ -8,7 +8,7 @@
  *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-package argo.staj;
+package argo;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
@@ -18,7 +18,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Iterator;
 
-import static argo.staj.JsonStreamElementMatchers.aJsonStreamElementMatching;
+import static argo.JsonStreamElementMatchers.aJsonStreamElementMatching;
 import static java.util.Arrays.asList;
 
 public final class ElementTrackingStajParserMatcher extends TypeSafeDiagnosingMatcher<Iterator<JsonStreamElement>> {