+++ src/testFixtures/java/argo/RoundTrippingStajParserMatcher.java
@@ -8,12 +8,10 @@
  *  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-package argo.staj;
+package argo;
 
-import argo.JsonParserScopeExpander;
 import argo.format.PrettyJsonBuilder;
 import argo.jdom.JsonNode;
-import argo.saj.InvalidSyntaxException;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeDiagnosingMatcher;